flic2lib

+++ flic2lib.xcframework/ios-arm64_x86_64-simulator/flic2lib.framework/Headers/FLICButtonEvent.h
@@ -95,13 +95,13 @@ typedef NS_ENUM(int8_t, FLICButtonEventGesture) {
 @property (nonatomic, readonly) double age;
 
 /// The X-axis accelerometer value at the time of the event.
-@property (nonatomic, readonly) int8_t x;
+@property (nonatomic, readonly) float x;
 
 /// The Y-axis accelerometer value at the time of the event.
-@property (nonatomic, readonly) int8_t y;
+@property (nonatomic, readonly) float y;
 
 /// The Z-axis accelerometer value at the time of the event.
-@property (nonatomic, readonly) int8_t z;
+@property (nonatomic, readonly) float z;
 
 /// Gesture, -1 if no gesture.
 @property (nonatomic, readonly) FLICButtonEventGesture gesture;
